style: 框架优化

+++ themis-trigger/src/main/java/com/achobeta/themis/trigger/meilisearch/http/LawDataFixController.java
@@ -124,7 +124,7 @@ public class LawDataFixController {
             }
             
             // 2. 创建分类映射
-            Map<Integer, LawCategory> categoryMap = new HashMap<>();
+            Map<Long, LawCategory> categoryMap = new HashMap<>();
             for (LawCategory category : categories) {
                 categoryMap.put(category.getLawId(), category);
             }
